+++ web/WEB-INF/classes/src/org/caisi/dao/FacilityMessageDAO.java
@@ -45,7 +45,7 @@ public class FacilityMessageDAO extends HibernateDaoSupport  {
 		this.getHibernateTemplate().saveOrUpdate(mesg);
 	}
 
-	public List getMessagesByFacilityId(Long facilityId) {
+	public List getMessagesByFacilityId(Integer facilityId) {
 		return this.getHibernateTemplate().find("from FacilityMessage fm where facilityId=? order by fm.expiry_date desc", facilityId);
 	}
 }
